backend/requirements.txt
pymongo==4.6.1
python-jose==3.3.0
python-dotenv==1.0.1
pytest==8.0.2
pytest-cov==4.1.0
flask==2.3.3
flask-cors==4.0.0 